% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_create_beautier_cmd.R
\name{lum_create_beautier_cmd}
\alias{lum_create_beautier_cmd}
\title{Create the \code{beautier} command from the user interface}
\usage{
lum_create_beautier_cmd(ui)
}
\arguments{
\item{ui}{the user interface, as created by \link{lum_create_ui}}
}
\description{
Create the \code{beautier} command from the user interface
}
