% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_create_beautier_cmd_from_ui.R
\name{lum_create_beautier_cmd_from_ui}
\alias{lum_create_beautier_cmd_from_ui}
\title{Create the \code{beautier} command from the user interface}
\usage{
lum_create_beautier_cmd_from_ui(ui, as_file)
}
\arguments{
\item{ui}{the user interface, as created by \link{lum_create_ui}}

\item{as_file}{if TRUE, show \code{create_beast_input_file},
else show \code{create_beast_input}}
}
\description{
Create the \code{beautier} command from the user interface
}
\author{
Richel J.C. Bilderbeek
}
