% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_create_clock_model_arg.R
\name{lum_create_clock_model_arg}
\alias{lum_create_clock_model_arg}
\title{Create the \code{clock_model} argument}
\usage{
lum_create_clock_model_arg(clock_model)
}
\arguments{
\item{clock_model}{the clock model, as text}
}
\description{
Create the \code{clock_model} argument
}
\author{
Richèl J.C. Bilderbeek
}
