% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_create_cmd.R
\name{lum_create_cmd}
\alias{lum_create_cmd}
\title{Create the \code{beautier} command from the user interface}
\usage{
lum_create_cmd(input_filename = "test.fas",
  output_filename = "beast2.xml", subst_model = "JC69",
  clock_model = "Strict Clock", tree_prior = "Yule Model",
  chain_length = 1e+07, store_every = -1, as_file = FALSE)
}
\arguments{
\item{subst_model}{name of the substitution model}

\item{clock_model}{name of the clock model}

\item{tree_prior}{name of the tree prior}

\item{chain_length}{MCMC chain length}

\item{store_every}{MCMC store interval}

\item{as_file}{if TRUE, show \code{create_beast_input_file},
else show \code{create_beast_input}}

\item{filename}{the name of the FASTA filename}
}
\description{
Create the \code{beautier} command from the user interface
}
\author{
Richel J.C. Bilderbeek
}
