% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_create_cmd_from_ui.R
\name{lum_create_cmd_from_ui}
\alias{lum_create_cmd_from_ui}
\title{Create the \code{beautier} command from the user interface}
\usage{
lum_create_cmd_from_ui(ui)
}
\arguments{
\item{ui}{the user interface, as created by \link{lum_create_ui}}
}
\description{
Create the \code{beautier} command from the user interface
}
\author{
Richel J.C. Bilderbeek
}
