% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_create_input_filename_arg.R
\name{lum_create_input_filename_arg}
\alias{lum_create_input_filename_arg}
\title{Create the \code{input_filename} argument}
\usage{
lum_create_input_filename_arg(filename)
}
\arguments{
\item{filename}{name of the file}
}
\description{
Create the \code{input_filename} argument
}
\author{
Richel J.C. Bilderbeek
}
