% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_create_output_filename_arg.R
\name{lum_create_output_filename_arg}
\alias{lum_create_output_filename_arg}
\title{Create the \code{output_filename} argument}
\usage{
lum_create_output_filename_arg(output_filename)
}
\arguments{
\item{output_filename}{the name of the to-be-created BEAST2 file}
}
\description{
Create the \code{output_filename} argument
}
\author{
Richèl J.C. Bilderbeek
}
