% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_create_server.R
\name{lum_create_server}
\alias{lum_create_server}
\title{Define server logic}
\usage{
lum_create_server(input, output)
}
\arguments{
\item{input}{the current user interface,
as created by \link{lum_create_ui}}

\item{output}{the new, changed user interface}
}
\description{
Define server logic
}
