% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_create_tree_prior_arg.R
\name{lum_create_tree_prior_arg}
\alias{lum_create_tree_prior_arg}
\title{Create the \code{tree_prior} argument}
\usage{
lum_create_tree_prior_arg(tree_prior)
}
\description{
Create the \code{tree_prior} argument
}
\author{
Richel J.C. Bilderbeek
}
