% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_create_xml.R
\name{lum_create_xml}
\alias{lum_create_xml}
\title{Show the XML file created by the displayed beautier command}
\usage{
lum_create_xml(beautier_cmd)
}
\arguments{
\item{beautier_cmd}{the \code{beautier} command}
}
\description{
Show the XML file created by the displayed beautier command
}
\author{
Richèl J.C. Bilderbeek
}
