% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_run.R
\name{lum_run}
\alias{lum_run}
\title{Run \code{lumier} Shiny application.}
\usage{
lum_run()
}
\description{
Run \code{lumier} Shiny application.
}
\seealso{
The user interface is created by \link{lum_create_ui}.
  The Shiny server is created by \link{lum_create_server}.
}
\author{
Richel J.C. Bilderbeek
}
