% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_create_mcmc_arg.R
\name{lum_create_mcmc_arg}
\alias{lum_create_mcmc_arg}
\title{Create the \code{beautier}'s \code{mcmc} argument}
\usage{
lum_create_mcmc_arg(ui)
}
\description{
Create the \code{beautier}'s \code{mcmc} argument
}
