% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_create_mcmc_arg.R
\name{lum_create_mcmc_arg}
\alias{lum_create_mcmc_arg}
\title{Create the \code{mcmc} argument}
\usage{
lum_create_mcmc_arg(chain_length = "10000000", store_every = -1)
}
\arguments{
\item{chain_length}{MCMC chain length}

\item{store_every}{MCMC store interval}
}
\description{
Create the \code{mcmc} argument
}
\author{
Richèl J.C. Bilderbeek
}
