% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lum_create_site_model_arg.R
\name{lum_create_site_model_arg}
\alias{lum_create_site_model_arg}
\title{Create the \code{site_model} argument}
\usage{
lum_create_site_model_arg(subst_model = "JC69",
  gamma_site_model = beautier::create_gamma_site_model())
}
\arguments{
\item{subst_model}{name of the substitution model}

\item{gamma_site_model}{a gamma_site_model,
as created by \link[beautier]{create_gamma_site_model}}
}
\description{
Create the \code{site_model} argument
}
\author{
Richel J.C. Bilderbeek
}
